+++ merlict_c89/mliMedium_json.h
@@ -1,0 +1,13 @@
+/* Copyright 2018-2020 Sebastian Achim Mueller */
+#ifndef MERLICT_C89_MLIMEDIUM_JSON_H_
+#define MERLICT_C89_MLIMEDIUM_JSON_H_
+
+#include "mliMedium.h"
+#include "mliJson.h"
+
+int mliMedium_malloc_from_json_str(struct mliMedium *med, const char *json_str);
+int mliMedium_malloc_from_json_token(
+        struct mliMedium *med,
+        const struct mliJson *json,
+        const uint64_t token);
+#endif
